Fix an import statement to comply with pyinstaller
When trying to package an application that contains VTK imports with pyinstaller, usually there is a lot of trouble to make it work. The previous version of this file always fails, while the proposed modification solves the problem.

If this (https://github.com/pyvista/pyvista-support/issues/167#issuecomment-671920473) hidden imports are made and the above change is implemented, then everything (at least in my use case) works.